smtp rework
