+++ contributions/api/utils/adminutils.py
@@ -15,10 +15,6 @@
 import controllers.configs as cfg
 import utils.mongoutils as mongoutils
 
-import smtplib
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
-
 """
 check if the logged in user is a superuser
 """
@@ -53,102 +49,3 @@ def check_if_reviewer(login_id):
             return True
 
     return False
-
-def establish_smtp_connection():
-    """
-    Method to establish SMTP connection
-    Args: None
-    Returns:
-        connection (obj): SMTP object
-        smtp_error (str): Error string
-        smtp_code (str): Error code
-    """
-    mail_from = cfg.SENDER_EMAIL
-    password = cfg.SENDER_EMAIL_PASSWORD
-    connection = smtplib.SMTP(host=cfg.SMTP_HOST, port=cfg.SMTP_PORT)
-    try:
-        connection.ehlo()
-    except smtplib.SMTPHeloError as ex:
-        return None, ex.smtp_error, ex.smtp_code
-    try:
-        connection.starttls()
-    except smtplib.SMTPConnectError as ex:
-        return None, ex.smtp_error, ex.smtp_code
-    try:
-        connection.login(mail_from, password)
-    except smtplib.SMTPAuthenticationError as ex:
-        return None, ex.smtp_error, ex.smtp_code
-    return connection, '', ''
-
-
-def test_smtp_connection(connection):
-    """
-    Method to test the SMTP connection
-    Args:
-        connection (obj): SMTP object
-    Returns:
-        (bool): True if connection is open, False if not
-    """
-    try:
-        status = connection.noop()[0]
-    except smtplib.SMTPServerDisconnected as ex:
-        status = -1
-    return True if status == 250 else False
-
-
-def quit_smtp_connection(connection):
-    """
-    Method to quit the SMTP connection
-    Args:
-        connection (obj): SMTP object
-    Returns:
-        (bool): True if connection is closed successfully, False if not
-        smtp_error (str): Error string if failure, empty string if success
-        smtp_code (str): Error code if failure, empty string if success
-    """
-    # quit connection if connection exists
-    if test_smtp_connection(connection):
-        try:
-            connection.quit()
-            return True, ' ', ' '
-        except smtplib.SMTPResponseException as ex:
-            return False, ex.smtp_error, ex.smtp_code
-        except smtplib.SMTPException as ex:
-            return False, ex.strerror, ex.errno
-
-
-def send_email(mail_to, subject, message, connection):
-    """
-    Method to send email to a user and print if success or failure
-    Args:
-        mail_to (str): email id of recipient
-        subject (str): subject of the email
-        message (str): message of the email
-        connection (obj): SMTP object
-    Returns:
-        (bool): True if success, False if failure
-        (str): Error code is failure, empty string if success
-        (str): Error message if failure, empty string if success
-    """
-    mail_subject = subject
-    mail_body = message
-    mail_from = cfg.SENDER_EMAIL
-
-    mimemsg = MIMEMultipart()
-    mimemsg['From'] = mail_from
-    mimemsg['To'] = mail_to
-    mimemsg['Subject'] = mail_subject
-    mimemsg.attach(MIMEText(mail_body, 'plain'))
-
-    if not test_smtp_connection(connection):
-        connection, error_str, error_code = establish_smtp_connection()
-        if error_code:
-            return False, error_str, error_code
-    try:
-        connection.send_message(mimemsg)
-    except smtplib.SMTPResponseException as ex:
-        return False, ex.smtp_error, ex.smtp_code
-    except smtplib.SMTPException as ex:
-        return False, ex.strerror, ex.errno
-    return True, ' ', ' '
-
